backend/requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-multipart==0.0.9
Pillow==10.4.0
numpy==1.26.4
torch==2.3.1
torchvision==0.18.1
requests==2.32.3
